requirements.txt
fastapi~=0.111.0
SQLAlchemy~=2.0.30
python-dotenv~=1.0.1